Darkmode added, logo changed (#50)
Created Dark Mode theme, which is enabled if the users system settings
are set to Dark. Created in Collaboration with other group for
consistent styling.
Edited logo. Logo now mainly says "Image To Map", unless accessed by a
domain in an array, where it will display as "PDF To Map". Array is to
be edited later.

---------

Co-authored-by: Lars Henrik Råkil <89202778+HenrikRaakil@users.noreply.github.com>