Implement Map Tile Generation and Serve Tiles via Mapbox Requests (#69)
### Major changes include:

#### Map Tile Generation and Serving:
- Added functionality to generate map tiles dynamically based on
requests from Mapbox. This allows us to serve custom tiles to the Mapbox
map component.

#### Dependency Adjustments:
- Introduced `rio_tiler` as a new dependency to facilitate efficient
tile generation. This library provides a robust set of tools for working
with raster data and will be integral to our tile-serving backend.

### Testing:
- Tested the tile generation endpoint with a variety of map zoom levels
and coordinates to ensure reliability and performance. (Needs more
testing)
      - 
- Ensured that the `OverlayView` component updates correctly with
varying project IDs and coordinates.